+++ Runtime/Interaction/Interactables/XRGrabInteractable.cs
@@ -1940,20 +1940,23 @@ namespace UnityEngine.XR.Interaction.Toolkit.Interactables
         /// <seealso cref="Grab"/>
         protected virtual void Drop()
         {
-            if (m_RetainTransformParent && m_OriginalSceneParent != null && !m_OriginalSceneParent.gameObject.activeInHierarchy)
+            if (m_RetainTransformParent && m_OriginalSceneParent != null)
             {
+                if (!m_OriginalSceneParent.gameObject.activeInHierarchy)
+                {
 #if UNITY_EDITOR
-                // Suppress the warning when exiting Play mode to avoid confusing the user
-                var exitingPlayMode = UnityEditor.EditorApplication.isPlaying && !UnityEditor.EditorApplication.isPlayingOrWillChangePlaymode;
+                    // Suppress the warning when exiting Play mode to avoid confusing the user
+                    var exitingPlayMode = UnityEditor.EditorApplication.isPlaying && !UnityEditor.EditorApplication.isPlayingOrWillChangePlaymode;
 #else
-                var exitingPlayMode = false;
+                    var exitingPlayMode = false;
 #endif
-                if (!exitingPlayMode)
-                    Debug.LogWarning("Retain Transform Parent is set to true, and has a non-null Original Scene Parent. " +
-                        "However, the old parent is deactivated so we are choosing not to re-parent upon dropping.", this);
+                    if (!exitingPlayMode)
+                        Debug.LogWarning("Retain Transform Parent is set to true, and has a non-null Original Scene Parent. " +
+                            "However, the old parent is deactivated so we are choosing not to re-parent upon dropping.", this);
+                }
+                else if (gameObject.activeInHierarchy)
+                    transform.SetParent(m_OriginalSceneParent);
             }
-            else if (m_RetainTransformParent && gameObject.activeInHierarchy)
-                transform.SetParent(m_OriginalSceneParent);
 
             SetupRigidbodyDrop(m_Rigidbody);